Add ShapezSteamModPackageManager
By the time this is being committed it is already obsolete. The original idea
was collaborating with the community on the loader. However, concerns about a
copy stability, effort to create a custom solution and the added difficulties
in supporting arbitrary loaders in the game made us go in a different direction